Create image_editing.py
This file proposes a solution to create a polygon on an image then generate a mask. Then uses Stable Diffusion model in Azure ML to edit the image in the polygon area just created.

It includes 5 ways to create a polygon. One way is to use prompt with gpt4-o to specify the polygon area.